+++ _examples/prompt_default/main.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"os/user"
 
-	"github.com/sohomdatta1/promptui"
+	"github.com/NozomiNetworks/promptui"
 )
 
 func main() {